requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
ping3==5.1.5
psutil==7.2.2
python-dotenv==1.2.2
requests==2.32.5
urllib3==2.6.3
python-telegram-bot==21.6